Updates the SplashScreenLayout
Updates the Splash Screen Layout as per the mockups.

Also fixes the issue of Splash Screen remaining open in background even after switching activity.

What needs to be changed?
The resource named ic_launcher with the app icon needs to be added to the project.

The splash screen will automatically reflect the changed app logo from the `ic_launcher` resource.